% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_mapping_functions.R
\name{color_reassign}
\alias{color_reassign}
\title{Reassign the microshades colors to different groups}
\usage{
color_reassign(cdf, group_assignment, color_assignment, group_level = "Phylum")
}
\arguments{
\item{cdf}{data.frame containing the color key}

\item{group_assignment}{string vector of taxonomic groups}

\item{color_assignment}{sting vector of corresponding color assignment}

\item{group_level}{string indiating the taxonomic level of group_assignment}
}
\value{
cdf data.frame
}
\description{
To customize the color order in the plot, use this function to directly assign
groups a specific color
}
\examples{
new_cdf <- color_reassign(cdf,
                          group_assignment = c("Firmicutes", "Actinobacteria"),
                          color_assignment = c("micro_cvd_blue", "micro_cvd_purple"))

}
