% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_mapping_functions.R
\name{create_color_dfs}
\alias{create_color_dfs}
\title{Generates abundance sorted data frame and a color palette data frame}
\usage{
create_color_dfs(
  mdf,
  selected_groups = c("Proteobacteria", "Actinobacteria", "Bacteroidetes",
    "Firmicutes"),
  top_n_subgroups = 4,
  group_level = "Phylum",
  subgroup_level = "Genus",
  cvd = FALSE,
  top_orientation = FALSE
)
}
\arguments{
\item{mdf}{data.frame, melted data frame containing microbiome data}

\item{selected_groups}{list of groups in group_level taxomomy to specify and color in plot.
The vector order is the stacking order. "Other" is always on the top of the stack,
but then the rest will follow. The default is "Proteobacteria", "Actinobacteria",
"Bacteroidetes", "Firmicutes". "Firmicutes" is on the bottom of the stack.}

\item{top_n_subgroups}{integer number of top subgroups to show for each selected group
the max is 4 top subgroups}

\item{group_level}{string of larger taxonomic group}

\item{subgroup_level}{string of smaller taxonomic group}

\item{cvd}{logical, determines which palette to use,
color vision deficent (microshades_cvd_palettes) or microshades_palettes}

\item{top_orientation}{logical most abundant shades oriented at the top of the stack
otherwise, most abundant shades are bottom oriented}
}
\value{
list
\itemize{
\item{"mdf"}{ melted data frame to pass into ggplot2}
\item{"cdf"}{ mapping to be used in manual color filling}
}
}
\description{
Use \code{create_color_dfs()} to create a specialized data frames containing color
and abundance organized microbiome data.
}
\details{
The top group categories are user specified through the \code{selected_groups} parameter,
and top subgroup categories are dynamically generated based on abundance.
For the top group, the categories not in \code{selected_groups} will be changed to
"Other". The \code{top_n_subgroups} will be determined for each selected group.

Notes:
\itemize{
\item In SILVA 138, some phylum names are different and you should consider
passing in the vector
\code{c("Proteobacteria", "Actinobacteriota", "Bacteroidota", "Firmicutes")}
}
}
\examples{
library(phyloseq)
data(GlobalPatterns)

mdf <- prep_mdf(GlobalPatterns)

colorframe <- create_color_dfs(mdf)
colorframe <- create_color_dfs(mdf, selected_groups = c("Proteobacteria", "Actinobacteriota", "Bacteroidota", "Firmicutes"))
colorframe <- create_color_dfs(mdf, cvd = TRUE)
}
