% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_legend.R
\name{custom_legend}
\alias{custom_legend}
\title{This function creates a custom legend that is organized by same group classifications}
\usage{
custom_legend(
  mdf,
  cdf,
  group_level = "Phylum",
  subgroup_level = "Genus",
  legend_key_size = 0.4,
  legend_text_size = 10
)
}
\arguments{
\item{mdf}{data.frame melted dataframe}

\item{cdf}{data.frame color dataframe}

\item{group_level}{string of larger taxonomic group}

\item{subgroup_level}{string of smaller taxonomic group}

\item{legend_key_size}{numeric determines overall size of the legend keys}

\item{legend_text_size}{integer determines size of legend text}
}
\value{
complete_legend ggplot
}
\description{
This function creates a custom legend that is organized by same group classifications
}
\examples{
legend_new <- custom_legend(mdf_GP, cdf_GP)


}
