% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_mapping_functions.R
\name{default_hex}
\alias{default_hex}
\title{Return default hex colors}
\usage{
default_hex(n_groups = 5, cvd = FALSE)
}
\arguments{
\item{n_groups}{integer of number of selected groups}

\item{cvd}{logical Color Vision Deficent Friendly palette useage}
}
\value{
data.frame
}
\description{
Returns a built-in data frame with hex code.
}
\examples{
# Get hex codes for 5 groups
hex_df <- default_hex()

# Get hex codes for 3 groups with CVD palette
hex_df <- default_hex(3, TRUE)
}
\keyword{internal}
