% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_mapping_functions.R
\name{extend_group}
\alias{extend_group}
\title{Extends the number of subgroups shown for one particular group by adding additional colors.}
\usage{
extend_group(
  mdf,
  cdf,
  group_level,
  subgroup_level,
  group_name,
  existing_palette,
  new_palette,
  n_add = 5,
  light = TRUE
)
}
\arguments{
\item{mdf}{melted data frame with microshades group processing}

\item{cdf}{color data frame}

\item{group_level}{string of larger taxonomic group}

\item{subgroup_level}{string of smaller taxonomic group}

\item{group_name}{name of the group to extend the palette for}

\item{existing_palette}{name of current palette of group to extend}

\item{new_palette}{name of new palette to add for extending the particular group's colors}

\item{n_add}{number of colors to add}

\item{light}{orientation of colors added; will make a difference if n_add is less than
the number of colors in the new palette}
}
\value{
list
\itemize{
\item{"mdf"}{new mdf with reclassified groups that include the group exention}
\item{"cdf"}{new cdf with reclassified groups that include the group exention}
}
}
\description{
Extends the number of subgroups shown for one particular group by adding additional colors.
}
\examples{

updated_objs <- extend_group(mdf, cdf, "Phylum", "Genus", "Firmicutes", "micro_purple", "micro_cvd_purple")

updated_objs$mdf
updated_objs$cdf

}
