% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_legend.R
\name{individual_legend}
\alias{individual_legend}
\title{Creates a small custom legend for one group variable}
\usage{
individual_legend(
  mdf,
  cdf,
  group_name,
  col_name_group = "Top_Phylum",
  col_name_subgroup = "Top_Genus",
  x = "Sample",
  y = "Abundance",
  legend_key_size = 0.4,
  legend_text_size = 10
)
}
\arguments{
\item{mdf}{data.frame melted dataframe}

\item{cdf}{data.frame color dataframe}

\item{group_name}{name of the larger taxonomic group to extract small legend from}

\item{col_name_group}{string column name of larger taxonomic group}

\item{col_name_subgroup}{string column nameof smaller taxonomic group}

\item{legend_key_size}{numeric determines overall size of the legend keys}

\item{legend_text_size}{integer determines size of legend text}
}
\value{
legend ggplot
}
\description{
Create a custom legend for one individual group, with the subgroups below
}
\keyword{internal}
