% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_mapping_functions.R
\name{match_cdf}
\alias{match_cdf}
\title{Apply the color factoring from one cdf to a different melted data frame}
\usage{
match_cdf(mdf, mdf_group, group_level = "Phylum", subgroup_level = "Genus")
}
\arguments{
\item{mdf}{data.frame, melted data frame to apply legend to}

\item{mdf_group}{data.frame, melted data frame to use legend from}

\item{group_level}{string of larger taxonomic group}

\item{subgroup_level}{string of smaller taxonomic group}
}
\value{
data.frame, melted df with ordered factor that are consistent with the legend
}
\description{
Now both melted dataframes will contain the same color mapping information.
This can be useful if you want to make sure that different graphs have consistent legends
}
\examples{
mdf_to_plot <- match_cdf(mdf, mdf_group)

}
