% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_mapping_functions.R
\name{match_color_df}
\alias{match_color_df}
\title{Apply the color factoring from one color df to a different melted phyloseq object.}
\usage{
match_color_df(
  mdf,
  mdf_group,
  selected_groups = c("Proteobacteria", "Actinobacteria", "Bacteroidetes",
    "Firmicutes"),
  group_level = "Phylum",
  subgroup_level = "Genus"
)
}
\arguments{
\item{mdf}{data.frame, melted data frame to apply legend to}

\item{mdf_group}{data.frame, melted data frame to use legend from}

\item{selected_groups}{list. "Other" is always on the top of the stack,
but then the rest will follow, with Firmicutes as the base of the stack}

\item{group_level}{string of larger taxonomic group}

\item{subgroup_level}{string of smaller taxonomic group}
}
\value{
data.frame, melted df with ordered factor that are consistent with the legend
}
\description{
Now both melted dataframes will contain the same color mapping information
}
\examples{
mdf_to_plot <- match_color_df(mdf, mdf_group)

}
