% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{microshades_cvd_palettes}
\alias{microshades_cvd_palettes}
\title{List of CVD palettes}
\format{
An object of class \code{list} of length 6.
}
\usage{
microshades_cvd_palettes
}
\description{
Use \code{\link{microshades_palette}} to construct palettes of desired length.
}
\examples{
microshades_cvd_palettes
microshades_palette("micro_cvd_gray")
microshades_palette("micro_cvd_green")
microshades_palette("micro_cvd_orange")
microshades_palette("micro_cvd_blue")
microshades_palette("micro_cvd_turquoise")
microshades_palette("micro_cvd_purple")

}
\keyword{datasets}
