% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_contributions}
\alias{plot_contributions}
\title{Creates median and mean abundance barplots, and boxplots for microshades taxa contribution plots keeping
colors consistent in the microshades stacked barplots}
\usage{
plot_contributions(mdf, cdf, col_name, data_match, short_name = NULL)
}
\arguments{
\item{mdf}{melted data frame with microshades group processing}

\item{cdf}{color data frame}

\item{col_name}{name of column to subset from}

\item{data_match}{specific data to subset match}

\item{short_name}{alternate group name in both the mdf and cdf, note the group column still must exist}
}
\value{
list
\itemize{
\item{"box"}{ boxplot}
\item{"mean"}{ barplot of mean abundance}
\item{"median"}{barplot of median abundance}
}
}
\description{
Creates median and mean abundance barplots, and boxplots for microshades taxa contribution plots keeping
colors consistent in the microshades stacked barplots
}
\examples{

plots <- plot_contributions(mdf, cdf,  "dmm", 1, short_name = "group_short")

plots$box
plots$mean
plots$median

plots <- plot_contributions(mdf, cdf,  "sex", "female")

}
