% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_mapping_functions.R
\name{plot_microshades}
\alias{plot_microshades}
\title{Plot the mdf created with the microshades package}
\usage{
plot_microshades(
  mdf_group,
  color_df,
  group_label = "Phylum Genus",
  x = "Sample",
  y = "Abundance"
)
}
\arguments{
\item{mdf_group}{data.frame, melted data frame processed by microshades functions}

\item{color_df}{data.frame containing the color key}

\item{group_label}{string of smaller taxonomic group}

\item{x}{string x-axis element}

\item{y}{string y-axis element}
}
\value{
A \code{\link{ggplot}}2 plot.
}
\description{
Plot the mdf created with the microshades package
}
\examples{
plot <- plot_microshades(mdf_group, color_df)
}
