% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_mapping_functions.R
\name{prep_mdf}
\alias{prep_mdf}
\title{Prepare melted phyloseq object to map colors to}
\usage{
prep_mdf(
  ps,
  subgroup_level = "Genus",
  as_relative_abundance = TRUE,
  remove_na = FALSE
)
}
\arguments{
\item{ps}{phyloseq-class object}

\item{subgroup_level}{string of smaller taxonomic group}

\item{as_relative_abundance}{transform counts to relative abundance}

\item{remove_na}{remove NA values during taxa agglomeration}
}
\value{
data.frame, a melted phyloseq object from \code{psmelt()}
}
\description{
This function agglomerates a phyloseq object at a specified level and transforms counts to relative abundance.
This function also melts the phyloseq object into a melted data frame which is used to apply microshades colors and organization to.
}
\details{
Notes:
\itemize{
\item This normalizes the phyloseq object to relative abundance
\item This agglomerates to the smaller taxonomic group \code{subgroup_level}
}
}
\examples{
library(phyloseq)
library(speedyseq)
data(GlobalPatterns)

# Use defaults
mdf <- prep_mdf(GlobalPatterns)

# Subgroup as "Family"
mdf_fam <- prep_mdf(GlobalPatterns, subgroup_level = "Family")
}
