% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_mapping_functions.R
\name{reorder_samples_by}
\alias{reorder_samples_by}
\title{Reorder the samples by abundance of user selected subgroup taxa}
\usage{
reorder_samples_by(
  mdf_group,
  order = "Lactobacillus",
  subgroup_level = "Genus"
)
}
\arguments{
\item{mdf_group}{data.frame, melted data frame}

\item{order}{string of subgroup to reorder by}

\item{subgroup_level}{string of smaller taxonomic group}
}
\value{
data.frame, mdf dataframe ready to plot
}
\description{
Reorder the samples by abundance of user selected subgroup taxa
}
\examples{
mdf_to_plot <- reorder_samples_by(mdf_group)
}
