% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_legend.R
\name{custom_legend}
\alias{custom_legend}
\title{Create a custom legend}
\usage{
custom_legend(
  mdf,
  cdf,
  group_level = "Phylum",
  subgroup_level = "Genus",
  x = "Sample",
  y = "Abundance",
  legend_key_size = 0.4,
  legend_text_size = 10,
  legend_orientation = "vertical"
)
}
\arguments{
\item{mdf}{data.frame melted dataframe}

\item{cdf}{data.frame color dataframe}

\item{group_level}{string of larger taxonomic group}

\item{subgroup_level}{string of smaller taxonomic group}

\item{x}{x-value plotted}

\item{y}{y-value plotted}

\item{legend_key_size}{numeric determines overall size of the legend keys}

\item{legend_text_size}{integer determines size of legend text}

\item{legend_orientation}{string determines vertical or horizontal orientation of legend}
}
\value{
complete_legend ggplot
}
\description{
A customized legend sections the group taxonomy as headers
and lists the subgroup taxonomies below each group header, for all the groups.
}
\examples{
legend_new <- custom_legend(mdf_GP, cdf_GP)


}
