% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{microshades_cvd_palettes}
\alias{microshades_cvd_palettes}
\title{List of CVD palettes}
\format{
An object of class \code{list} of length 6.
}
\usage{
microshades_cvd_palettes
}
\description{
Use \code{\link{microshades_cvd_palette}} to construct palettes of desired length.
}
\keyword{datasets}
