% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{microshades_palette}
\alias{microshades_palette}
\title{Microshades sequential color shading palettes}
\usage{
microshades_palette(name, n, lightest = TRUE)
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{micro_gray}, \code{micro_brown},  \code{micro_green},
\code{micro_orange}, \code{micro_blue},  \code{micro_purple}
\code{micro_cvd_gray}, \code{micro_cvd_brown},  \code{micro_cvd_green},
\code{micro_cvd_orange}, \code{micro_cvd_blue},  \code{micro_cvd_purple}}

\item{n}{Number of colors desired. Palettes contain 5 shades
If omitted, uses all colours.}

\item{lightest}{logical, if user requests n < 5 it will reutrn n lightest shades if TRUE.
If FALSE, it will return n darkest shades}
}
\value{
A vector of colours.
}
\description{
These palettes use a base color paired with lighter versions of the base color
}
\examples{
microshades_palette("micro_gray")
microshades_palette("micro_brown")
microshades_palette("micro_cvd_green")
microshades_palette("micro_orange", 3)

}
\keyword{colors}
