% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{microshades_palettes}
\alias{microshades_palettes}
\title{List of palettes}
\format{
An object of class \code{list} of length 6.
}
\usage{
microshades_palettes
}
\description{
Use \code{\link{microshades_palette}} to construct palettes of desired length.
}
\examples{
microshades_palettes
microshades_palette("micro_gray")
microshades_palette("micro_brown")
microshades_palette("micro_green")
microshades_palette("micro_orange")
microshades_palette("micro_blue")
microshades_palette("micro_purple")

}
\keyword{datasets}
