% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_microshades}
\alias{plot_microshades}
\title{Plot the mdf created with the microshades package}
\usage{
plot_microshades(
  mdf_group,
  cdf,
  group_label = "Phylum Genus",
  x = "Sample",
  y = "Abundance"
)
}
\arguments{
\item{mdf_group}{data.frame, melted data frame processed by microshades functions}

\item{cdf}{data.frame containing the color key}

\item{group_label}{string of smaller taxonomic group}

\item{x}{string x-axis element}

\item{y}{string y-axis element}
}
\value{
A \code{\link{ggplot}}2 plot.
}
\description{
Creates a stacked abundance ggplot with color abundance organization
}
\examples{
library(phyloseq)
data(GlobalPatterns)

mdf <- prep_mdf(GlobalPatterns)

color_obj <- create_color_dfs(mdf)

mdf_group <- color_obj$mdf
cdf <- color_obj$cdf

plot <- plot_microshades(mdf_group, cdf)
}
