% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_mapping_functions.R
\name{prep_mdf}
\alias{prep_mdf}
\title{Prepare melted phyloseq object to map colors to}
\usage{
prep_mdf(ps, subgroup_level = "Genus")
}
\arguments{
\item{ps}{phyloseq-class object}

\item{subgroup_level}{string of smaller taxonomic group}
}
\value{
data.frame, a melted phyloseq object from \code{psmelt()}
}
\description{
This function prepares a melted phyloseq object to map colors to top phyla
and genera.
}
\details{
Notes:
\itemize{
\item This normalizes the phyloseq object to relative abundance
\item This agglomerates to the smaller taxonomic group \code{subgroup_level}
}
}
\examples{
data(GlobalPatterns)

# Use defaults
mdf <- prep_mdf(GlobalPatterns)

# Subgroup as "Family"
mdf_fam <- prep_mdf(GlobalPatterns, subgroup_level = "Family")
}
