% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_mapping_functions.R
\name{reorder_samples_by}
\alias{reorder_samples_by}
\title{Reorder the samples or stacked group levels by abundance}
\usage{
reorder_samples_by(
  mdf_group,
  cdf,
  order = "NA",
  group_level = "Phylum",
  subgroup_level = "Genus",
  sample_variable = "Sample",
  sink_abundant_groups = TRUE
)
}
\arguments{
\item{mdf_group}{data.frame, data frame containf microbiome data}

\item{cdf}{data.frame containing the color key}

\item{order}{string of subgroup to reorder by}

\item{group_level}{string of larger taxonomic group}

\item{subgroup_level}{string of smaller taxonomic group}

\item{sample_variable}{sample variable to reorder (x- axis component for plot)}

\item{sink_abundant_groups}{logical reorder the phylum groups so the most abundant is the bottom group}
}
\value{
list
\itemize{
\item{"mdf"}{ reordered melted data frame, ready for plotting}
\item{"cdf"}{ reordered manual color filling according to new order}
}
}
\description{
This function will reorder the user selected subgroup taxa based on abundance, and can also
reorder the stacked groups levels based on abundance, using \code{sink_abundant_groups}
}
\examples{
library(phyloseq)
data(GlobalPatterns)

mdf <- prep_mdf(GlobalPatterns)

color_obj <- create_color_dfs(mdf)

mdf_group <- color_obj$mdf
cdf <- color_obj$cdf

mdf_new <- reorder_samples_by(mdf_group, cdf)
mdf_new <- reorder_samples_by(mdf_group, cdf, order = "Bacteroides")
}
